Factor out helper functions.
